# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ЖерБақылау API",
    "description": "Цифровой мониторинг земель: панель инспектора, «Народный контроль» и статусы заявлений.\n\nОшибки возвращаются в формате `{error: {code, message, details}}`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/public/acts/{act_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Verify a printed act by its QR code",
        "operationId": "verify_act_api_v1_public_acts__act_id__get",
        "parameters": [
          {
            "name": "act_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Act Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActPublic"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/inspections/{token}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Owner: open a photo-report request",
        "operationId": "open_inspection_api_v1_public_inspections__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionPublic"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "public"
        ],
        "summary": "Owner: submit live camera photos with the device position (anti-fraud checks run server-side)",
        "operationId": "submit_inspection_api_v1_public_inspections__token__post",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_inspection_api_v1_public_inspections__token__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionSubmitted"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActPublic": {
        "properties": {
          "number": {
            "type": "string",
            "title": "Number"
          },
          "issued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Issued At"
          },
          "cadastral_number": {
            "type": "string",
            "title": "Cadastral Number"
          },
          "parcel_status": {
            "type": "string",
            "title": "Parcel Status"
          },
          "lang": {
            "type": "string",
            "title": "Lang"
          },
          "issued_by": {
            "type": "string",
            "title": "Issued By"
          },
          "pdf_sha256": {
            "type": "string",
            "title": "Pdf Sha256"
          },
          "photo_count": {
            "type": "integer",
            "title": "Photo Count"
          },
          "chain_at_issue": {
            "$ref": "#/components/schemas/ChainCheck",
            "description": "Audit log unchanged since the act was issued"
          }
        },
        "type": "object",
        "required": [
          "number",
          "issued_at",
          "cadastral_number",
          "parcel_status",
          "lang",
          "issued_by",
          "pdf_sha256",
          "photo_count",
          "chain_at_issue"
        ],
        "title": "ActPublic",
        "description": "What anyone scanning the QR code on a printed act can see (no personal data)."
      },
...
      "Body_submit_inspection_api_v1_public_inspections__token__post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "1–5 JPEG photos captured by the camera"
          },
          "captured_at": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Captured At",
            "description": "ISO time of each capture, same order as files"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "accuracy": {
            "type": "number",
            "title": "Accuracy",
            "description": "GPS accuracy, metres"
          },
          "declared_use": {
            "$ref": "#/components/schemas/DeclaredUse"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "files",
          "captured_at",
          "lat",
          "lon",
          "accuracy",
          "declared_use"
        ],
        "title": "Body_submit_inspection_api_v1_public_inspections__token__post"
      },
...
      "ChainCheck": {
        "properties": {
          "intact": {
            "type": "boolean",
            "title": "Intact"
          },
          "length": {
            "type": "integer",
            "title": "Length"
          },
          "head": {
            "type": "string",
            "title": "Head"
          },
          "broken_at_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broken At Id"
          }
        },
        "type": "object",
        "required": [
          "intact",
          "length",
          "head"
        ],
        "title": "ChainCheck"
      },
...
      "DeclaredUse": {
        "type": "string",
        "enum": [
          "CULTIVATED",
          "BUILDING",
          "FALLOW",
          "CLEANED",
          "OTHER"
        ],
        "title": "DeclaredUse"
      },
...
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "InspectionPublic": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "status": {
            "$ref": "#/components/schemas/InspectionStatus"
          },
          "cadastral_number": {
            "type": "string",
            "title": "Cadastral Number"
          },
          "address_ru": {
            "type": "string",
            "title": "Address Ru"
          },
          "address_kk": {
            "type": "string",
            "title": "Address Kk"
          },
          "purpose": {
            "$ref": "#/components/schemas/ParcelPurpose"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "centroid": {
            "$ref": "#/components/schemas/LngLat"
          },
          "parcel": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parcel",
            "description": "Parcel boundary (GeoJSON) to show where to stand"
          }
        },
        "type": "object",
        "required": [
          "code",
          "status",
          "cadastral_number",
          "address_ru",
          "address_kk",
          "purpose",
          "due_at",
          "note",
          "centroid",
          "parcel"
        ],
        "title": "InspectionPublic",
        "description": "What the owner sees on the one-time link (no personal data of anyone)."
      },
...
      "InspectionStatus": {
        "type": "string",
        "enum": [
          "REQUESTED",
          "SUBMITTED",
          "ACCEPTED",
          "REJECTED",
          "EXPIRED"
        ],
        "title": "InspectionStatus"
      },
      "InspectionSubmitted": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "status": {
            "$ref": "#/components/schemas/InspectionStatus"
          },
          "photos": {
            "type": "integer",
            "title": "Photos"
          }
        },
        "type": "object",
        "required": [
          "code",
          "status",
          "photos"
        ],
        "title": "InspectionSubmitted"
      },
...
      "LngLat": {
        "properties": {
          "lon": {
            "type": "number",
            "title": "Lon"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          }
        },
        "type": "object",
        "required": [
          "lon",
          "lat"
        ],
        "title": "LngLat"
      },
...
      "ParcelPurpose": {
        "type": "string",
        "enum": [
          "IZHS",
          "AGRICULTURE",
          "COMMERCIAL",
          "INDUSTRIAL",
          "LPH"
        ],
        "title": "ParcelPurpose"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Supabase access token (or local dev token)",
        "scheme": "bearer"
      }
    }
  }
}